# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"SearchClickRequest": {
				"properties": {
					"searchLogId": {
						"type": "number",
						"format": "double"
					},
					"searchLogItemId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"searchLogId",
					"searchLogItemId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RecommendationResponse": {
				"properties": {
					"recommendationLogId": {
						"type": "number",
						"format": "double"
					},
					"data": {
						"items": {},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"pageSize": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"pageSize",
							"page",
							"total"
						],
						"type": "object"
					},
					"processingTimeMs": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"pagination",
					"processingTimeMs"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RecommendErrorMessage": {
				"type": "string",
				"enum": [
					"Page must be >= 1",
					"PageSize must be between 1 and 50",
					"No recommendation history yet. Browse some posts first!",
					"Recommendation service is temporarily unavailable",
					"recommendationLogId and recommendationLogItemId are required",
					"Failed to get recommendations",
					"Failed to log click",
					""
				],
				"description": "Union type for all possible Recommendation error messages"
			},
			"RecommendErrorResponse": {
				"description": "Error response structure for Recommendation API (Legacy format)",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"$ref": "#/components/schemas/RecommendErrorMessage"
					}
				},
				"required": [
					"success",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RecommendationClickRequest": {
				"properties": {
					"recommendationLogId": {
						"type": "number",
						"format": "double"
					},
					"recommendationLogItemId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"recommendationLogId",
					"recommendationLogItemId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "tro-tot-vn-be",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "lamdx4"
		}
	},
	"paths": {
...
		"/search/click": {
			"post": {
				"operationId": "LogClick",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Log user click on search result",
				"tags": [
					"Search"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SearchClickRequest"
							}
						}
					}
				}
			}
		},
		"/recommend": {
			"get": {
				"operationId": "GetRecommendations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RecommendationResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RecommendErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RecommendErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get personalized recommendations for authenticated customer",
				"tags": [
					"Recommend"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "logId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/recommend/click": {
			"post": {
				"operationId": "LogClick",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RecommendErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RecommendErrorResponse"
								}
							}
						}
					}
				},
				"description": "Log user click on recommendation",
				"tags": [
					"Recommend"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RecommendationClickRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}